requirements.txt
Django==2.0
psycopg2-binary==2.7.4
djangorestframework==3.7.7



